requirements.txt
annotated-types==0.6.0
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.2
Flask-Cors==4.0.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.24.2
pandas==2.2.0
pydantic==2.6.1
pydantic_core==2.16.2
python-dateutil==2.8.2
python-jobspy==1.1.42
pytz==2024.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
tls-client==1.0.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
Werkzeug==3.0.1
